feat: reorg step 1, game class
