+++ Game.h
@@ -1,0 +1,51 @@
+#pragma once
+
+#include "helper.h"
+#include <string>
+#include <vector>
+#include "vulkanControl.h"
+#include "windowControl.h"
+
+class Game {
+public:
+    Game();
+    ~Game();
+
+    void run();
+
+private:
+   
+    GLFWwindow* window = nullptr;
+    WindowControl* windowController = nullptr;
+    VulkanControl* vulkanController = nullptr;
+
+    std::vector<Vertex> vertices1;
+    std::vector<uint32_t> indices1;
+
+    std::vector<Vertex> vertices2;
+    std::vector<uint32_t> indices2;
+
+    uint32_t currentFrame = 0;
+
+    bool framebufferResized = false;
+
+    void initWindow();
+    void initVulkan();
+    void mainLoop();
+    void cleanup();
+
+    void drawFrame();
+
+    void processInput();
+
+    static void framebufferResizeCallback(GLFWwindow* window, int width, int height) {
+        auto app = reinterpret_cast<Game*>(glfwGetWindowUserPointer(window));
+        app->framebufferResized = true;
+    }
+
+    bool hasStencilComponent(VkFormat format) {
+        return format == VK_FORMAT_D32_SFLOAT_S8_UINT || format == VK_FORMAT_D24_UNORM_S8_UINT;
+    }
+
+    void loadModel(std::string modelPath, std::vector<Vertex>& destVer, std::vector<uint32_t>& destIndices);
+};
